+++ src/olink/core/types.h
@@ -23,26 +23,9 @@
 */
 #pragma once
 
-#include <string>
-#include <stdio.h>
-#include <stdlib.h>
-#include <map>
-#include <list>
+#include "olink_common.h"
 #include "nlohmann/json.hpp"
-
-#if defined _WIN32 || defined __CYGWIN__
-#ifdef __GNUC__
-  #define OLINK_EXPORT __attribute__ ((dllexport))
-#else
-  #define OLINK_EXPORT __declspec(dllexport)
-#endif
-#else
-  #if __GNUC__ >= 4
-    #define OLINK_EXPORT __attribute__ ((visibility ("default")))
-  #else
-    #define OLINK_EXPORT
-  #endif
-#endif
+#include <string>
 
 namespace ApiGear { namespace ObjectLink {
